[GH-90] move test_process unit tests to test_wrapper
